chore: Slight redesign of `ListItem` "API"
- Instead of having separate `labelText` & `labelTextKey`, we have a single `labelText` which we pass into the `t()` function.
  - We know that passing a non-translation key will result in the passed string being returned.
- The `LeftElement`/RightElement` props has been replaced with `Leading` & `Trailing`.
